README.md
We have a number of in house offers, custom landing pages and direct relationships with advertisers which also allow us to be the most competitive when it comes to payouts and performance for the traffic verticals we cater to. Our top categories are Movies, Sports, Games, Sweeps, Books and Downloads. Our highly optimized smartlinks will also ensure you get the best paying CPA and highest performing offer for complete world wide coverage. We support payments methods such as Paypal, Wire, Payoneer, Bitcoin, ePayments